requirements.txt
# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2
seaborn==0.13.2

# MLflow
mlflow==2.8.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2

# Utilities
python-multipart==0.0.6
requests==2.31.0
opencensus-ext-azure==1.1.15 